+++ packages/location-utils/types/index.d.ts
@@ -1,5 +1,2 @@
 export declare function ensureBeginsWith(str: string | undefined | null, prefix: string): string;
-export declare function completePathname(pathname?: string): string;
-export declare function completeHash(hash?: string): string;
-export declare function completeQuery(query?: string): string;
 export declare function stripPrefix(str: string, prefix: string): string;