+++ packages/location-utils/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+- Remove `complete___` exports.
+
 ## 2.0.0-beta.13
 
 - Remove `stripBase` function.